Embarassing humongous commit. It:
Moves towards putting user-facing functionality into the browser
(chrome for now). Some highlights:

The launcher is no longer assumed to be an independant window, but just
opens as a browser tab. Hence window-managing of that is gone.

The notifier removed in favour of a chrome-exension. The extension
connects to RefudeServices and listens for notifications. When such
arrive, they are shown, using the browsers notifications.

Some cleanup - the pub-sub functionality of the watch endpoint now in a
library, and all connections have a publisher associated.